Change support function patterns to drastically improve performance
Changed the support function patterns so that one pattern runs first to determine if the function is in a user-defined namespace, and if not we assume it's a built-in function. I also slightly restructured how the patterns are used. With these changes I was able to eliminate a lot of redundancy and significantly improve performance.

Using Allan's gtm tool I was able to estimate the performance difference from this change by running test-cases.php through the old and new grammars. Using the grammar prior to any PHP 5.3 changes (79ef7ab5d) I get an average of 1.4s @ 5616 bytes/s. The new grammar is averaging 0.6s @ 14153 bytes/s.