Improve/simplify escaping rules for single-quoted regexps
The rules have been generalized so that anything can be escaped (in the regex) where previously we only supported escaping known special characters. I think the former is better, since there shouldn’t be an issue with over escaping things in a regex, whereas not matching all escape sequences can easily lead to incorrect parsing.

The previous version of this grammar missed some edge-cases and accepted double-backslash in front of an apostrophe (as a way to escape the quote).

Closes #59.