Allow <?php ?> snippet to work in any document
This works anywhere in a PHP document and only at the very start of any other document. Along with the recent firstLineMatch changes in TextMate this means using the snippet on the first line of an untitled document will automatically set the document language to PHP.

Due to this change the ⌃> shortcut has been removed.